+++ ToDoChangeLog.md
@@ -1,8 +1,7 @@
 # Stuff to Change on this Development Branch
 Merge Star, Planet, and Satellite Body(from https://github.com/rademonster/3S/blob/master/system.py) classes into one class; Body 
 
-# Phase 1
-Merge Star, Planet, and Satellite classes into Body
+# Phase 1, COMPLETED
 
 # Phase 2
 Ensure Body compatability with System(from https://github.com/rademonster/3S/blob/master/system.py)
